Fixed parsing and stream handling

+++ src/main/java/com/engineersbox/conduit/schema/metric/MetricValueType.java
@@ -4,7 +4,7 @@ public enum MetricValueType {
     STRING(String.class),
     FLOAT(float.class),
     DOUBLE(double.class),
-    INT(int.class),
+    INT(long.class),
     BOOLEAN(boolean.class),
     CONTAINER(null);
